V-2.60.00 - new gcad-Version Microsoft/MSYS2 MINGW_x86_64 (Win10, 64-bit) - bugfixes ..

+++ src/gr/ut_GLU.c
@@ -2338,7 +2338,8 @@ static int ptOff;
   GLT_ppa[GLT_ppa_ind].typ  = Typ_Color;
   GLT_ppa[GLT_ppa_ind].form = Typ_Int4;
   GLT_ppa[GLT_ppa_ind].siz  = 1;
-  GLT_ppa[GLT_ppa_ind].data = (void*)(*((long*)&GL_actCol));
+  GLT_ppa[GLT_ppa_ind].data = (void*)(*((long long*)&GL_actCol));   // OGX_SET_COLOR
+  // GLT_ppa[GLT_ppa_ind].data = (void*)(*((long*)&GL_actCol));
   // OGX_SET_COLOR(&GLT_ppa[GLT_ppa_ind], &GL_actCol);
 
   return 0;
